# protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Robot Local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "summary": "Health",
        "operationId": "health_v1_models_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/speech/recognize": {
      "post": {
        "summary": "Recognize",
        "operationId": "recognize_v1_speech_recognize_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_v1_speech_recognize_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/turns": {
      "post": {
        "summary": "Turn",
        "operationId": "turn_v1_turns_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Turn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/speech/reply": {
      "post": {
        "summary": "Reply",
        "operationId": "reply_v1_speech_reply_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Speak"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_v1_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_recognize_v1_speech_recognize_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_v1_speech_recognize_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Register": {
        "properties": {
          "invite": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "Invite"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite",
          "name"
        ],
        "title": "Register"
      },
...
      "Speak": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 600,
            "minLength": 1,
            "title": "Text"
          },
          "voice": {
            "$ref": "#/components/schemas/Voice"
          },
          "story": {
            "type": "boolean",
            "title": "Story",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Speak"
      },
      "Turn": {
        "properties": {
          "sessionId": {
            "type": "string",
            "maxLength": 80,
            "minLength": 16,
            "title": "Sessionid"
          },
          "text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text"
          },
          "image": {
            "type": "string",
            "maxLength": 1500000,
            "title": "Image",
            "default": ""
          },
          "imageAgeMs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Imageagems",
            "default": 0
          },
          "visualRequest": {
            "type": "boolean",
            "title": "Visualrequest",
            "default": false
          },
          "currentResourceId": {
            "type": "string",
            "maxLength": 80,
            "title": "Currentresourceid",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessionId",
          "text"
        ],
        "title": "Turn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Voice": {
        "properties": {
          "zh": {
            "type": "string",
            "maxLength": 80,
            "title": "Zh",
            "default": "default"
          },
          "en": {
            "type": "string",
            "maxLength": 80,
            "title": "En",
            "default": "default"
          },
          "story": {
            "type": "string",
            "maxLength": 80,
            "title": "Story",
            "default": "default"
          },
          "speed": {
            "type": "number",
            "maximum": 1.3,
            "minimum": 0.7,
            "title": "Speed",
            "default": 1.0
          },
          "volume": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Volume",
            "default": 0.5
          },
          "style": {
            "type": "string",
            "enum": [
              "neutral",
              "gentle",
              "cheerful",
              "storytelling",
              "soothing"
            ],
            "title": "Style",
            "default": "gentle"
          },
          "storyStyle": {
            "type": "string",
            "enum": [
              "neutral",
              "gentle",
              "cheerful",
              "storytelling",
              "soothing"
            ],
            "title": "Storystyle",
            "default": "storytelling"
          },
          "instruction": {
            "type": "string",
            "maxLength": 200,
            "title": "Instruction",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Voice"
      }
    }
  }
}